separate planners into directory
